postgres: remove use of sqlx
This commit rewrites the "postgres" packages to only use pgxpool for
their database access. This reduces the number of database handles
passed around and makes idioms for queries consistent.

Some uses of the "distlock" package in tests was mocked out, under the
reasoning that all tests have a database constructed for them anyway,
via the integration helper.